+++ launchpad.php
@@ -35,9 +35,9 @@ add_action( 'activated_plugin', function(){
  */
 require_once 'global/setup.php';
 
-require_once 'make-admin/index.php';
+require_once 'vendor/make-admin/index.php';
 
-require_once 'admin/settings.php';
+require_once 'admin-ui/settings.php';
 
 // // Back-end
 // require_once 'admin/admin-field.php';
